@@ -8,8 +8,8 @@
     "@testing-library/user-event": "^13.5.0",
     "html-entities": "^2.3.2",
     "nanoid": "^3.3.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-scripts": "5.0.0",
     "web-vitals": "^2.1.4"
   },
